refactor: move entities into core folder
