+++ api/handler/v1beta1/grpc.go
@@ -6,11 +6,11 @@ import (
 	"strings"
 
 	guardianv1beta1 "github.com/odpf/guardian/api/proto/odpf/guardian/v1beta1"
-	"github.com/odpf/guardian/appeal"
+	"github.com/odpf/guardian/core/appeal"
+	"github.com/odpf/guardian/core/policy"
+	"github.com/odpf/guardian/core/provider"
+	"github.com/odpf/guardian/core/resource"
 	"github.com/odpf/guardian/domain"
-	"github.com/odpf/guardian/policy"
-	"github.com/odpf/guardian/provider"
-	"github.com/odpf/guardian/resource"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
