+++ app/server.go
@@ -11,8 +11,11 @@ import (
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	handlerv1beta1 "github.com/odpf/guardian/api/handler/v1beta1"
 	guardianv1beta1 "github.com/odpf/guardian/api/proto/odpf/guardian/v1beta1"
-	"github.com/odpf/guardian/appeal"
-	"github.com/odpf/guardian/approval"
+	"github.com/odpf/guardian/core/appeal"
+	"github.com/odpf/guardian/core/approval"
+	"github.com/odpf/guardian/core/policy"
+	"github.com/odpf/guardian/core/provider"
+	"github.com/odpf/guardian/core/resource"
 	"github.com/odpf/guardian/domain"
 	"github.com/odpf/guardian/iam"
 	"github.com/odpf/guardian/internal/crypto"
@@ -25,9 +28,6 @@ import (
 	"github.com/odpf/guardian/plugins/providers/grafana"
 	"github.com/odpf/guardian/plugins/providers/metabase"
 	"github.com/odpf/guardian/plugins/providers/tableau"
-	"github.com/odpf/guardian/policy"
-	"github.com/odpf/guardian/provider"
-	"github.com/odpf/guardian/resource"
 	"github.com/odpf/guardian/store"
 	"github.com/odpf/salt/log"
 	"golang.org/x/net/http2"
