+++ core/appeal/errors.go
