+++ core/approval/repository_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/DATA-DOG/go-sqlmock"
-	"github.com/odpf/guardian/approval"
+	"github.com/odpf/guardian/core/approval"
 	"github.com/odpf/guardian/domain"
 	"github.com/odpf/guardian/mocks"
 	"github.com/odpf/guardian/utils"
