+++ core/resource/errors.go
